updating web app to use subject id from core lib + some code cleanup and bug fixes
