+++ packages/webapp/src/pages/Select/AggregateStatistics/AggregateStatistics.hooks.ts
@@ -38,6 +38,7 @@ export function useGetAggregateStatistics(): (
 						}),
 						{
 							delimiter: sensitiveContent.delimiter,
+							subjectId: sensitiveContent.subjectId,
 							useColumns: headers,
 							sensitiveZeros: sensitiveContent.headers
 								.filter(h => h.hasSensitiveZeros)
