--- packages/webapp/src/pages/Select/DataSelect/hooks/commands/index.ts
@@ -1,11 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-export * from './useColumnsWithZeros.js'
-export * from './useOnSensitiveZeroCheckChange.js'
-export * from './useOnUseColumnCheckToggle.js'
-export * from './useSensitiveZerosCommand.js'
-export * from './useSubjectIdCommand.js'
-export * from './useVisibleColumnNames.js'
-export * from './useVisibleColumnsCommand.js'
