--- packages/webapp/src/pages/Select/DataSelect/hooks/commands/useColumnsWithZeros.ts
@@ -1,20 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-import { useMemo } from 'react'
-
-import type { ICsvContent } from '~models'
-
-export function useColumnsWithZeros(
-	content: ICsvContent,
-): string[] | undefined {
-	return useMemo(
-		() =>
-			content.columnsWithZeros
-				?.map(i => content.headers[i])
-				.filter(h => h.use)
-				.map(h => h.name),
-		[content],
-	)
-}
