--- packages/webapp/src/pages/Select/DataSelect/hooks/index.ts
@@ -1,7 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-export * from './commands/index.js'
-export * from './useIsSubjectIdValid.js'
-export * from './useSensitiveTableCommands.js'
