+++ packages/webapp/src/pages/Select/SelectCommands/SensitiveZeros/SensitiveZeros.hooks.tsx
@@ -2,10 +2,23 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import { useCallback } from 'react'
+import { useCallback, useMemo } from 'react'
 import type { SetterOrUpdater } from 'recoil'
 
 import type { ICsvContent } from '~models'
+
+export function useColumnsWithZeros(
+	content: ICsvContent,
+): string[] | undefined {
+	return useMemo(
+		() =>
+			content.columnsWithZeros
+				?.map(i => content.headers[i])
+				.filter(h => h.use)
+				.map(h => h.name),
+		[content],
+	)
+}
 
 export function useOnSensitiveZeroCheckToggle(
 	setter: SetterOrUpdater<ICsvContent>,
