+++ packages/webapp/src/pages/hooks/useCanRun.ts
@@ -6,14 +6,11 @@ import { useMemo } from 'react'
 
 import { useIsProcessingValue, useSensitiveContentValue } from '~states'
 
-import { useIsSubjectIdValid } from '../Select/DataSelect/hooks/index.js'
-
 export function useCanRun(): boolean {
 	const content = useSensitiveContentValue()
 	const isProcessing = useIsProcessingValue()
-	const isValid = useIsSubjectIdValid(content)
 
 	return useMemo(() => {
-		return content.table.totalRows() > 0 && !isProcessing && isValid === true
-	}, [content, isProcessing, isValid])
+		return content.table.totalRows() > 0 && !isProcessing
+	}, [content, isProcessing])
 }
